Convert money balances to decimal

+++ WalletApp.Domain/Entities/Account.cs
@@ -2,8 +2,8 @@ namespace WalletApp.Domain.Entities
 {
     public class Account : Entity
     {
-        public float CardBalance { get; set; }
-        public float AvailableBalance { get; set; }
+        public decimal CardBalance { get; set; }
+        public decimal AvailableBalance { get; set; }
         public string? NoPaymentDueMessage { get; set; }
         public int DailyPoints { get; set; }
     }
