+++ WalletApp.Domain/Entities/Transaction.cs
@@ -3,7 +3,7 @@ namespace WalletApp.Domain.Entities
     public class Transaction : Entity
     {
         public TransactionType TransactionType { get; set; }
-        public float Sum { get; set; }
+        public decimal Sum { get; set; }
         public string Title { get; set; } = string.Empty;
         public string? Description { get; set; }
         public DateTime Date { get; set; }
